+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * deps: basic-auth@~1.0.1
   * deps: depd@~1.0.1
   * deps: on-finished@~2.2.1
     - Fix `isFinished(req)` when data buffered